Moved Input code to dedicated class
