+++ Runtime/CAVE/CAVEInputManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e56f7be2cfe66f247b7105e1d20cf604
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
