+++ Runtime/CaveTouchInput.cs
@@ -136,7 +136,7 @@ public partial class @CaveTouchInput: IInputActionCollection2, IDisposable
                     ""path"": ""<Pointer>/delta"",
                     ""interactions"": """",
                     ""processors"": """",
-                    ""groups"": """",
+                    ""groups"": "";Cave Demo Control Scheme"",
                     ""action"": ""Drag"",
                     ""isComposite"": false,
                     ""isPartOfComposite"": false
