+++ Runtime/CaveTouchInput.inputactions
@@ -50,7 +50,7 @@
                     "path": "<Pointer>/delta",
                     "interactions": "",
                     "processors": "",
-                    "groups": "",
+                    "groups": ";Cave Demo Control Scheme",
                     "action": "Drag",
                     "isComposite": false,
                     "isPartOfComposite": false
